add script. script has errors.
Note the interesting error when using this approach. Also note the
resources being used with this approach in 3-ld_calc.out. which() seems
to be one function that is not compatible with really long vectors, as
its implemented in /src/include/Rinlinedfuns.h:137. Check
implementation. Could I build my own C++ implementation?